+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * A `pre_activitypub_get_upload_baseurl` filter
+* Fediverse Preview on post-overview page
 * GitHub action to enforce Changelog updates.
 
 ### Improved
 
 * Outsource Constants to a separate file
+* Better handling of `readme.txt` and `README.md`
 
 ## 4.2.1 - 2024-11-20